requirements.txt
setuptools==57.0.0
requests==2.28.1
rel==0.4.7
websocket-client==1.3.3
urllib3==1.26.11